split launch methods out of agent
