+++ src/radical/pilot/agent/__init__.py
@@ -4,5 +4,6 @@ __license__   = "MIT"
 
 
 from .rm        import RM
+from .lm        import LaunchMethod as LM
 from .scheduler import Scheduler
