+++ src/radical/pilot/agent/lm/aprun.py
@@ -1,0 +1,56 @@
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .base import LaunchMethod
+
+
+# ==============================================================================
+#
+class APRun(LaunchMethod):
+
+    # --------------------------------------------------------------------------
+    #
+    def __init__(self, cfg, logger):
+
+        LaunchMethod.__init__(self, cfg, logger)
+
+
+    # --------------------------------------------------------------------------
+    #
+    def _configure(self):
+        # aprun: job launcher for Cray systems
+        self.launch_command= self._which('aprun')
+
+        # TODO: ensure that only one concurrent aprun per node is executed!
+
+
+    # --------------------------------------------------------------------------
+    #
+    def construct_command(self, cu, launch_script_hop):
+
+        opaque_slots = cu['opaque_slots']
+        cud          = cu['description']
+        task_exec    = cud['executable']
+        task_cores   = cud['cores']
+        task_mpi     = cud['mpi']
+        task_args    = cud.get('arguments') or []
+        task_argstr  = self._create_arg_string(task_args)
+
+        if task_argstr:
+            task_command = "%s %s" % (task_exec, task_argstr)
+        else:
+            task_command = task_exec
+
+        if task_mpi:
+            pes = task_cores
+        else:
+            pes = 1
+        aprun_command = "%s -n %d %s" % (self.launch_command, pes, task_command)
+
+        return aprun_command, None
+
+
+# ------------------------------------------------------------------------------
+
