+++ src/radical/pilot/agent/lm/ccmrun.py
@@ -1,0 +1,48 @@
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .base import LaunchMethod
+
+# ==============================================================================
+#
+class CCMRun(LaunchMethod):
+
+    # --------------------------------------------------------------------------
+    #
+    def __init__(self, cfg, logger):
+
+        LaunchMethod.__init__(self, cfg, logger)
+
+
+    # --------------------------------------------------------------------------
+    #
+    def _configure(self):
+        # ccmrun: Cluster Compatibility Mode (CCM) job launcher for Cray systems
+        self.launch_command= self._which('ccmrun')
+
+
+    # --------------------------------------------------------------------------
+    #
+    def construct_command(self, cu, launch_script_hop):
+
+        opaque_slots = cu['opaque_slots']
+        cud          = cu['description']
+        task_exec    = cud['executable']
+        task_cores   = cud['cores']
+        task_args    = cud.get('arguments') or []
+        task_argstr  = self._create_arg_string(task_args)
+
+        if task_argstr:
+            task_command = "%s %s" % (task_exec, task_argstr)
+        else:
+            task_command = task_exec
+
+        ccmrun_command = "%s -n %d %s" % (self.launch_command, task_cores, task_command)
+
+        return ccmrun_command, None
+
+
+# ------------------------------------------------------------------------------
+
