+++ src/radical/pilot/agent/lm/dplace.py
@@ -1,0 +1,59 @@
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .base import LaunchMethod
+
+
+# ==============================================================================
+#
+class DPlace(LaunchMethod):
+
+    # --------------------------------------------------------------------------
+    #
+    def __init__(self, cfg, logger):
+
+        LaunchMethod.__init__(self, cfg, logger)
+
+
+    # --------------------------------------------------------------------------
+    #
+    def _configure(self):
+        # dplace: job launcher for SGI systems (e.g. on Blacklight)
+        self.launch_command = self._which('dplace')
+
+
+    # --------------------------------------------------------------------------
+    #
+    def construct_command(self, cu, launch_script_hop):
+
+        opaque_slots = cu['opaque_slots']
+        cud          = cu['description']
+        task_exec    = cud['executable']
+        task_cores   = cud['cores']
+        task_args    = cud.get('arguments') or []
+        task_argstr  = self._create_arg_string(task_args)
+
+        if 'task_offsets' not in opaque_slots :
+            raise RuntimeError('insufficient information to launch via %s: %s' \
+                    % (self.name, opaque_slots))
+
+        task_offsets = opaque_slots['task_offsets']
+
+        if task_argstr:
+            task_command = "%s %s" % (task_exec, task_argstr)
+        else:
+            task_command = task_exec
+
+        dplace_offset = task_offsets
+
+        dplace_command = "%s -c %d-%d %s" % (
+            self.launch_command, dplace_offset,
+            dplace_offset+task_cores-1, task_command)
+
+        return dplace_command, None
+
+
+# ------------------------------------------------------------------------------
+
