+++ src/radical/pilot/agent/lm/ibrun.py
@@ -1,0 +1,62 @@
+
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .base import LaunchMethod
+
+
+# ==============================================================================
+#
+class IBRun(LaunchMethod):
+    # NOTE: Don't think that with IBRUN it is possible to have
+    # processes != cores ...
+
+    # --------------------------------------------------------------------------
+    #
+    def __init__(self, cfg, logger):
+
+        LaunchMethod.__init__(self, cfg, logger)
+
+
+    # --------------------------------------------------------------------------
+    #
+    def _configure(self):
+        # ibrun: wrapper for mpirun at TACC
+        self.launch_command = self._which('ibrun')
+
+
+    # --------------------------------------------------------------------------
+    #
+    def construct_command(self, cu, launch_script_hop):
+
+        opaque_slots = cu['opaque_slots']
+        cud          = cu['description']
+        task_exec    = cud['executable']
+        task_cores   = cud['cores']
+        task_args    = cud.get('arguments') or []
+        task_argstr  = self._create_arg_string(task_args)
+
+        if not 'task_offsets' in opaque_slots:
+            raise RuntimeError('insufficient information to launch via %s: %s' \
+                    % (self.name, opaque_slots))
+
+        task_offsets = opaque_slots['task_offsets']
+
+        if task_argstr:
+            task_command = "%s %s" % (task_exec, task_argstr)
+        else:
+            task_command = task_exec
+
+        ibrun_offset = task_offsets
+
+        ibrun_command = "%s -n %s -o %d %s" % \
+                        (self.launch_command, task_cores,
+                         ibrun_offset, task_command)
+
+        return ibrun_command, None
+
+
+# ------------------------------------------------------------------------------
+
