+++ src/radical/pilot/agent/lm/mpirun.py
@@ -1,0 +1,65 @@
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .base import LaunchMethod
+
+
+# ==============================================================================
+#
+class MPIRun(LaunchMethod):
+
+    # --------------------------------------------------------------------------
+    #
+    def __init__(self, cfg, logger):
+
+        LaunchMethod.__init__(self, cfg, logger)
+
+
+    # --------------------------------------------------------------------------
+    #
+    def _configure(self):
+        self.launch_command = self._find_executable([
+            'mpirun',            # General case
+            'mpirun_rsh',        # Gordon @ SDSC
+            'mpirun-mpich-mp',   # Mac OSX MacPorts
+            'mpirun-openmpi-mp'  # Mac OSX MacPorts
+        ])
+
+
+    # --------------------------------------------------------------------------
+    #
+    def construct_command(self, cu, launch_script_hop):
+
+        opaque_slots = cu['opaque_slots']
+        cud          = cu['description']
+        task_exec    = cud['executable']
+        task_cores   = cud['cores']
+        task_args    = cud.get('arguments') or []
+        task_argstr  = self._create_arg_string(task_args)
+
+        if not 'task_slots' in opaque_slots:
+            raise RuntimeError('insufficient information to launch via %s: %s' \
+                    % (self.name, opaque_slots))
+
+        task_slots = opaque_slots['task_slots']
+
+        if task_argstr:
+            task_command = "%s %s" % (task_exec, task_argstr)
+        else:
+            task_command = task_exec
+
+        # Construct the hosts_string
+        hosts_string = ",".join([slot.split(':')[0] for slot in task_slots])
+
+        export_vars = ' '.join(['-x ' + var for var in self.EXPORT_ENV_VARIABLES if var in os.environ])
+
+        mpirun_command = "%s %s -np %s -host %s %s" % (
+            self.launch_command, export_vars, task_cores, hosts_string, task_command)
+
+        return mpirun_command, None
+
+
+# ------------------------------------------------------------------------------
+
