+++ src/radical/pilot/agent/lm/mpirun_ccmrun.py
@@ -1,0 +1,68 @@
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .base import LaunchMethod
+
+
+# ==============================================================================
+#
+class MPIRunCCMRun(LaunchMethod):
+    # TODO: This needs both mpirun and ccmrun
+
+    # --------------------------------------------------------------------------
+    #
+    def __init__(self, cfg, logger):
+
+        LaunchMethod.__init__(self, cfg, logger)
+
+
+    # --------------------------------------------------------------------------
+    #
+    def _configure(self):
+        # ccmrun: Cluster Compatibility Mode job launcher for Cray systems
+        self.launch_command= self._which('ccmrun')
+
+        self.mpirun_command = self._which('mpirun')
+        if not self.mpirun_command:
+            raise RuntimeError("mpirun not found!")
+
+
+    # --------------------------------------------------------------------------
+    #
+    def construct_command(self, cu, launch_script_hop):
+
+        opaque_slots = cu['opaque_slots']
+        cud          = cu['description']
+        task_exec    = cud['executable']
+        task_cores   = cud['cores']
+        task_args    = cud.get('arguments') or []
+        task_argstr  = self._create_arg_string(task_args)
+
+        if not 'task_slots' in opaque_slots:
+            raise RuntimeError('insufficient information to launch via %s: %s' \
+                    % (self.name, opaque_slots))
+
+        task_slots = opaque_slots['task_slots']
+
+        if task_argstr:
+            task_command = "%s %s" % (task_exec, task_argstr)
+        else:
+            task_command = task_exec
+
+        # Construct the hosts_string
+        # TODO: is there any use in using $HOME/.crayccm/ccm_nodelist.$JOBID?
+        hosts_string = ','.join([slot.split(':')[0] for slot in task_slots])
+        export_vars  = ' '.join(['-x ' + var for var in self.EXPORT_ENV_VARIABLES \
+                                             if  var in os.environ])
+
+        mpirun_ccmrun_command = "%s %s %s -np %d -host %s %s" % (
+            self.launch_command, self.mpirun_command, export_vars,
+            task_cores, hosts_string, task_command)
+
+        return mpirun_ccmrun_command, None
+
+
+# ------------------------------------------------------------------------------
+
