+++ src/radical/pilot/agent/lm/mpirun_dplace.py
@@ -1,0 +1,61 @@
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .base import LaunchMethod
+
+
+# ==============================================================================
+#
+class MPIRunDPlace(LaunchMethod):
+    # TODO: This needs both mpirun and dplace
+
+    # --------------------------------------------------------------------------
+    #
+    def __init__(self, cfg, logger):
+
+        LaunchMethod.__init__(self, cfg, logger)
+
+
+    # --------------------------------------------------------------------------
+    #
+    def _configure(self):
+        # dplace: job launcher for SGI systems (e.g. on Blacklight)
+        self.launch_command = self._which('dplace')
+        self.mpirun_command = self._which('mpirun')
+
+
+    # --------------------------------------------------------------------------
+    #
+    def construct_command(self, cu, launch_script_hop):
+
+        opaque_slots = cu['opaque_slots']
+        cud          = cu['description']
+        task_exec    = cud['executable']
+        task_cores   = cud['cores']
+        task_args    = cud.get('arguments') or []
+        task_argstr  = self._create_arg_string(task_args)
+
+        if not 'task_offsets' in opaque_slots:
+            raise RuntimeError('insufficient information to launch via %s: %s' \
+                    % (self.name, opaque_slots))
+
+        task_offsets = opaque_slots['task_offsets']
+
+        if task_argstr:
+            task_command = "%s %s" % (task_exec, task_argstr)
+        else:
+            task_command = task_exec
+
+        dplace_offset = task_offsets
+
+        mpirun_dplace_command = "%s -np %d %s -c %d-%d %s" % \
+            (self.mpirun_command, task_cores, self.launch_command,
+             dplace_offset, dplace_offset+task_cores-1, task_command)
+
+        return mpirun_dplace_command, None
+
+
+# ------------------------------------------------------------------------------
+
