+++ src/radical/pilot/agent/lm/poe.py
@@ -1,0 +1,72 @@
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .base import LaunchMethod
+
+
+# ==============================================================================
+#
+class POE(LaunchMethod):
+
+    # --------------------------------------------------------------------------
+    #
+    def __init__(self, cfg, logger):
+
+        LaunchMethod.__init__(self, cfg, logger)
+
+
+    # --------------------------------------------------------------------------
+    #
+    def _configure(self):
+        # poe: LSF specific wrapper for MPI (e.g. yellowstone)
+        self.launch_command = self._which('poe')
+
+
+    # --------------------------------------------------------------------------
+    #
+    def construct_command(self, cu, launch_script_hop):
+
+        opaque_slots = cu['opaque_slots']
+        cud          = cu['description']
+        task_exec    = cud['executable']
+        task_cores   = cud['cores']
+        task_args    = cud.get('arguments') or []
+        task_argstr  = self._create_arg_string(task_args)
+
+        if not 'task_slots' in opaque_slots:
+            raise RuntimeError('insufficient information to launch via %s: %s' \
+                    % (self.name, opaque_slots))
+
+        task_slots = opaque_slots['task_slots']
+
+        # Count slots per host in provided slots description.
+        hosts = {}
+        for slot in task_slots:
+            host = slot.split(':')[0]
+            if host not in hosts:
+                hosts[host] = 1
+            else:
+                hosts[host] += 1
+
+        # Create string with format: "hostX N host
+        hosts_string = ''
+        for host in hosts:
+            hosts_string += '%s %d ' % (host, hosts[host])
+
+        if task_argstr:
+            task_command = "%s %s" % (task_exec, task_argstr)
+        else:
+            task_command = task_exec
+
+        # Override the LSB_MCPU_HOSTS env variable as this is set by
+        # default to the size of the whole pilot.
+        poe_command = 'LSB_MCPU_HOSTS="%s" %s %s' % (
+            hosts_string, self.launch_command, task_command)
+
+        return poe_command, None
+
+
+# ------------------------------------------------------------------------------
+
