+++ src/radical/pilot/agent/radical-pilot-agent-multicore.py
@@ -201,23 +201,6 @@ git_ident = "$Id$"
 # ------------------------------------------------------------------------------
 # CONSTANTS
 #
-# 'enum' for unit launch method types
-LAUNCH_METHOD_APRUN         = 'APRUN'
-LAUNCH_METHOD_CCMRUN        = 'CCMRUN'
-LAUNCH_METHOD_DPLACE        = 'DPLACE'
-LAUNCH_METHOD_FORK          = 'FORK'
-LAUNCH_METHOD_IBRUN         = 'IBRUN'
-LAUNCH_METHOD_MPIEXEC       = 'MPIEXEC'
-LAUNCH_METHOD_MPIRUN_CCMRUN = 'MPIRUN_CCMRUN'
-LAUNCH_METHOD_MPIRUN_DPLACE = 'MPIRUN_DPLACE'
-LAUNCH_METHOD_MPIRUN        = 'MPIRUN'
-LAUNCH_METHOD_MPIRUN_RSH    = 'MPIRUN_RSH'
-LAUNCH_METHOD_ORTE          = 'ORTE'
-LAUNCH_METHOD_POE           = 'POE'
-LAUNCH_METHOD_RUNJOB        = 'RUNJOB'
-LAUNCH_METHOD_SSH           = 'SSH'
-LAUNCH_METHOD_YARN          = 'YARN'
-
 # 'enum' for pilot's unit spawner types
 SPAWNER_NAME_POPEN          = "POPEN"
 SPAWNER_NAME_SHELL          = "SHELL"
@@ -396,1631 +379,6 @@ def pilot_DONE(mongo_p=None, pilot_uid=None, logger=None, msg=None):
 
 
 
-# ==============================================================================
-#
-# Launch Methods
-#
-# ==============================================================================
-#
-class LaunchMethod(object):
-
-    # List of environment variables that designated Launch Methods should export
-    EXPORT_ENV_VARIABLES = [
-        'LD_LIBRARY_PATH',
-        'PATH',
-        'PYTHONPATH',
-        'PYTHON_DIR',
-        'RADICAL_PILOT_PROFILE'
-    ]
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        self.name = type(self).__name__
-        self._cfg = cfg
-        self._log = logger
-
-        # A per-launch_method list of environment to remove from the CU environment
-        self.env_removables = []
-
-        self.launch_command = None
-        self._configure()
-        # TODO: This doesn't make too much sense for LM's that use multiple
-        #       commands, perhaps this needs to move to per LM __init__.
-        if self.launch_command is None:
-            raise RuntimeError("Launch command not found for LaunchMethod '%s'" % self.name)
-
-        logger.info("Discovered launch command: '%s'.", self.launch_command)
-
-
-    # --------------------------------------------------------------------------
-    #
-    # This class-method creates the appropriate sub-class for the Launch Method.
-    #
-    @classmethod
-    def create(cls, name, cfg, logger):
-
-        # Make sure that we are the base-class!
-        if cls != LaunchMethod:
-            raise TypeError("LaunchMethod factory only available to base class!")
-
-        # In case of undefined LM just return None
-        if not name:
-            return None
-
-        try:
-            impl = {
-                LAUNCH_METHOD_APRUN         : LaunchMethodAPRUN,
-                LAUNCH_METHOD_CCMRUN        : LaunchMethodCCMRUN,
-                LAUNCH_METHOD_DPLACE        : LaunchMethodDPLACE,
-                LAUNCH_METHOD_FORK          : LaunchMethodFORK,
-                LAUNCH_METHOD_IBRUN         : LaunchMethodIBRUN,
-                LAUNCH_METHOD_MPIEXEC       : LaunchMethodMPIEXEC,
-                LAUNCH_METHOD_MPIRUN_CCMRUN : LaunchMethodMPIRUNCCMRUN,
-                LAUNCH_METHOD_MPIRUN_DPLACE : LaunchMethodMPIRUNDPLACE,
-                LAUNCH_METHOD_MPIRUN        : LaunchMethodMPIRUN,
-                LAUNCH_METHOD_MPIRUN_RSH    : LaunchMethodMPIRUNRSH,
-                LAUNCH_METHOD_ORTE          : LaunchMethodORTE,
-                LAUNCH_METHOD_POE           : LaunchMethodPOE,
-                LAUNCH_METHOD_RUNJOB        : LaunchMethodRUNJOB,
-                LAUNCH_METHOD_SSH           : LaunchMethodSSH,
-                LAUNCH_METHOD_YARN          : LaunchMethodYARN
-            }[name]
-            return impl(cfg, logger)
-
-        except KeyError:
-            logger.exception("LaunchMethod '%s' unknown or defunct" % name)
-
-        except Exception as e:
-            logger.exception("LaunchMethod cannot be used: %s!" % e)
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def lrms_config_hook(cls, name, cfg, lrms, logger):
-        """
-        This hook will allow the LRMS to perform launch methods specific
-        configuration steps.  The LRMS layer MUST ensure that this hook is
-        called exactly once (globally).  This will be a NOOP for LMs which do
-        not overload this method.  Exceptions fall through to the LRMS.
-        """
-
-        # Make sure that we are the base-class!
-        if cls != LaunchMethod:
-            raise TypeError("LaunchMethod config hook only available to base class!")
-
-        impl = {
-            LAUNCH_METHOD_FORK          : LaunchMethodFORK,
-            LAUNCH_METHOD_ORTE          : LaunchMethodORTE,
-            LAUNCH_METHOD_YARN          : LaunchMethodYARN
-        }.get(name)
-
-        if not impl:
-            logger.info('no LRMS config hook defined for LaunchMethod %s' % name)
-            return None
-
-        logger.info('call LRMS config hook for LaunchMethod %s: %s' % (name, impl))
-        return impl.lrms_config_hook(name, cfg, lrms, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def lrms_shutdown_hook(cls, name, cfg, lrms, lm_info, logger):
-        """
-        This hook is symmetric to the config hook above, and is called during
-        shutdown sequence, for the sake of freeing allocated resources.
-        """
-
-        # Make sure that we are the base-class!
-        if cls != LaunchMethod:
-            raise TypeError("LaunchMethod shutdown hook only available to base class!")
-
-        impl = {
-            LAUNCH_METHOD_ORTE          : LaunchMethodORTE,
-            LAUNCH_METHOD_YARN          : LaunchMethodYARN
-        }.get(name)
-
-        if not impl:
-            logger.info('no LRMS shutdown hook defined for LaunchMethod %s' % name)
-            return None
-
-        logger.info('call LRMS shutdown hook for LaunchMethod %s: %s' % (name, impl))
-        return impl.lrms_shutdown_hook(name, cfg, lrms, lm_info, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        raise NotImplementedError("_configure() not implemented for LaunchMethod: %s." % self.name)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-        raise NotImplementedError("construct_command() not implemented for LaunchMethod: %s." % self.name)
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def _find_executable(cls, names):
-        """Takes a (list of) name(s) and looks for an executable in the path.
-        """
-
-        if not isinstance(names, list):
-            names = [names]
-
-        for name in names:
-            ret = cls._which(name)
-            if ret is not None:
-                return ret
-
-        return None
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def _which(cls, program):
-        """Finds the location of an executable.
-        Taken from:
-        http://stackoverflow.com/questions/377017/test-if-executable-exists-in-python
-        """
-        # ----------------------------------------------------------------------
-        #
-        def is_exe(fpath):
-            return os.path.isfile(fpath) and os.access(fpath, os.X_OK)
-
-        fpath, _ = os.path.split(program)
-        if fpath:
-            if is_exe(program):
-                return program
-        else:
-            for path in os.environ["PATH"].split(os.pathsep):
-                exe_file = os.path.join(path, program)
-                if is_exe(exe_file):
-                    return exe_file
-        return None
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def _create_hostfile(cls, all_hosts, separator=' ', impaired=False):
-
-        # Open appropriately named temporary file
-        handle, filename = tempfile.mkstemp(prefix='rp_hostfile', dir=os.getcwd())
-
-        if not impaired:
-            #
-            # Write "hostN x\nhostM y\n" entries
-            #
-
-            # Create a {'host1': x, 'host2': y} dict
-            counter = collections.Counter(all_hosts)
-            # Convert it into an ordered dict,
-            # which hopefully resembles the original ordering
-            count_dict = collections.OrderedDict(sorted(counter.items(), key=lambda t: t[0]))
-
-            for (host, count) in count_dict.iteritems():
-                os.write(handle, '%s%s%d\n' % (host, separator, count))
-
-        else:
-            #
-            # Write "hostN\nhostM\n" entries
-            #
-            for host in all_hosts:
-                os.write(handle, '%s\n' % host)
-
-        # No longer need to write
-        os.close(handle)
-
-        # Return the filename, caller is responsible for cleaning up
-        return filename
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def _compress_hostlist(cls, all_hosts):
-
-        # Return gcd of a list of numbers
-        def gcd_list(l):
-            return reduce(fractions.gcd, l)
-
-        # Create a {'host1': x, 'host2': y} dict
-        count_dict = dict(collections.Counter(all_hosts))
-        # Find the gcd of the host counts
-        host_gcd = gcd_list(set(count_dict.values()))
-
-        # Divide the host counts by the gcd
-        for host in count_dict:
-            count_dict[host] /= host_gcd
-
-        # Recreate a list of hosts based on the normalized dict
-        hosts = []
-        [hosts.extend([host] * count)
-                for (host, count) in count_dict.iteritems()]
-        # Esthetically sort the list, as we lost ordering by moving to a dict/set
-        hosts.sort()
-
-        return hosts
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _create_arg_string(self, args):
-
-        # unit Arguments (if any)
-        arg_string = ''
-        if args:
-            for arg in args:
-                if not arg:
-                    # ignore empty args
-                    continue
-
-                arg = arg.replace('"', '\\"')    # Escape all double quotes
-                if arg[0] == arg[-1] == "'" :    # If a string is between outer single quotes,
-                    arg_string += '%s ' % arg    # ... pass it as is.
-                else:
-                    arg_string += '"%s" ' % arg  # Otherwise return between double quotes.
-
-        return arg_string
-
-
-
-# ==============================================================================
-#
-class LaunchMethodFORK(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # "Regular" tasks
-        self.launch_command = ''
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def lrms_config_hook(cls, name, cfg, lrms, logger):
-        return {'version_info': {
-            name: {'version': '0.42', 'version_detail': 'There is no spoon'}}}
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if task_argstr:
-            command = "%s %s" % (task_exec, task_argstr)
-        else:
-            command = task_exec
-
-        return command, None
-
-
-# ==============================================================================
-#
-class LaunchMethodMPIRUN(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        self.launch_command = self._find_executable([
-            'mpirun',            # General case
-            'mpirun_rsh',        # Gordon @ SDSC
-            'mpirun-mpich-mp',   # Mac OSX MacPorts
-            'mpirun-openmpi-mp'  # Mac OSX MacPorts
-        ])
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if not 'task_slots' in opaque_slots:
-            raise RuntimeError('insufficient information to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        task_slots = opaque_slots['task_slots']
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        # Construct the hosts_string
-        hosts_string = ",".join([slot.split(':')[0] for slot in task_slots])
-
-        export_vars = ' '.join(['-x ' + var for var in self.EXPORT_ENV_VARIABLES if var in os.environ])
-
-        mpirun_command = "%s %s -np %s -host %s %s" % (
-            self.launch_command, export_vars, task_cores, hosts_string, task_command)
-
-        return mpirun_command, None
-
-
-# ==============================================================================
-#
-class LaunchMethodSSH(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-        # Instruct the ExecWorkers to unset this environment variable.
-        # Otherwise this will break nested SSH with SHELL spawner, i.e. when
-        # both the sub-agent and CUs are started using SSH.
-        self.env_removables.extend(["RP_SPAWNER_HOP"])
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # Find ssh command
-        command = self._which('ssh')
-
-        if command is not None:
-
-            # Some MPI environments (e.g. SGE) put a link to rsh as "ssh" into
-            # the path.  We try to detect that and then use different arguments.
-            if os.path.islink(command):
-
-                target = os.path.realpath(command)
-
-                if os.path.basename(target) == 'rsh':
-                    self._log.info('Detected that "ssh" is a link to "rsh".')
-                    return target
-
-            command = '%s -o StrictHostKeyChecking=no' % command
-
-        self.launch_command = command
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if not 'task_slots' in opaque_slots:
-            raise RuntimeError('insufficient information to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        task_slots = opaque_slots['task_slots']
-
-        if not launch_script_hop :
-            raise ValueError ("LaunchMethodSSH.construct_command needs launch_script_hop!")
-
-        # Get the host of the first entry in the acquired slot
-        host = task_slots[0].split(':')[0]
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        # Pass configured and available environment variables to the remote shell
-        export_vars = ' '.join(['%s=%s' % (var, os.environ[var]) for var in self.EXPORT_ENV_VARIABLES if var in os.environ])
-
-        # Command line to execute launch script via ssh on host
-        ssh_hop_cmd = "%s %s %s %s" % (self.launch_command, host, export_vars, launch_script_hop)
-
-        # Special case, return a tuple that overrides the default command line.
-        return task_command, ssh_hop_cmd
-
-
-# ==============================================================================
-#
-class LaunchMethodMPIEXEC(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # mpiexec (e.g. on SuperMUC)
-        self.launch_command = self._find_executable([
-            'mpiexec',            # General case
-            'mpiexec-mpich-mp',   # Mac OSX MacPorts
-            'mpiexec-openmpi-mp'  # Mac OSX MacPorts
-        ])
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if not 'task_slots' in opaque_slots:
-            raise RuntimeError('insufficient information to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        task_slots = opaque_slots['task_slots']
-
-        # Extract all the hosts from the slots
-        all_hosts = [slot.split(':')[0] for slot in task_slots]
-
-        # Shorten the host list as much as possible
-        hosts = self._compress_hostlist(all_hosts)
-
-        # If we have a CU with many cores, and the compression didn't work
-        # out, we will create a hostfile and pass  that as an argument
-        # instead of the individual hosts
-        if len(hosts) > 42:
-
-            # Create a hostfile from the list of hosts
-            hostfile = self._create_hostfile(all_hosts, separator=':')
-            hosts_string = "-hostfile %s" % hostfile
-
-        else:
-
-            # Construct the hosts_string ('h1 h2 .. hN')
-            hosts_string = "-host "+ ",".join(hosts)
-
-        # Construct the executable and arguments
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        mpiexec_command = "%s -n %s %s %s" % (
-            self.launch_command, task_cores, hosts_string, task_command)
-
-        return mpiexec_command, None
-
-
-# ==============================================================================
-#
-class LaunchMethodAPRUN(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # aprun: job launcher for Cray systems
-        self.launch_command= self._which('aprun')
-
-        # TODO: ensure that only one concurrent aprun per node is executed!
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_mpi     = cud['mpi']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        if task_mpi:
-            pes = task_cores
-        else:
-            pes = 1
-        aprun_command = "%s -n %d %s" % (self.launch_command, pes, task_command)
-
-        return aprun_command, None
-
-
-
-# ==============================================================================
-#
-class LaunchMethodCCMRUN(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # ccmrun: Cluster Compatibility Mode (CCM) job launcher for Cray systems
-        self.launch_command= self._which('ccmrun')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        ccmrun_command = "%s -n %d %s" % (self.launch_command, task_cores, task_command)
-
-        return ccmrun_command, None
-
-
-
-# ==============================================================================
-#
-class LaunchMethodMPIRUNCCMRUN(LaunchMethod):
-    # TODO: This needs both mpirun and ccmrun
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # ccmrun: Cluster Compatibility Mode job launcher for Cray systems
-        self.launch_command= self._which('ccmrun')
-
-        self.mpirun_command = self._which('mpirun')
-        if not self.mpirun_command:
-            raise RuntimeError("mpirun not found!")
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if not 'task_slots' in opaque_slots:
-            raise RuntimeError('insufficient information to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        task_slots = opaque_slots['task_slots']
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        # Construct the hosts_string
-        # TODO: is there any use in using $HOME/.crayccm/ccm_nodelist.$JOBID?
-        hosts_string = ",".join([slot.split(':')[0] for slot in task_slots])
-
-        export_vars = ' '.join(['-x ' + var for var in self.EXPORT_ENV_VARIABLES if var in os.environ])
-
-        mpirun_ccmrun_command = "%s %s %s -np %d -host %s %s" % (
-            self.launch_command, self.mpirun_command, export_vars,
-            task_cores, hosts_string, task_command)
-
-        return mpirun_ccmrun_command, None
-
-
-
-# ==============================================================================
-#
-class LaunchMethodRUNJOB(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # runjob: job launcher for IBM BG/Q systems, e.g. Joule
-        self.launch_command= self._which('runjob')
-
-        raise NotImplementedError('RUNJOB LM needs to be decoupled from the scheduler/LRMS')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if  'cores_per_node'      not in opaque_slots or\
-            'loadl_bg_block'      not in opaque_slots or\
-            'sub_block_shape_str' not in opaque_slots or\
-            'corner_node'         not in opaque_slots :
-            raise RuntimeError('insufficient information to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        cores_per_node      = opaque_slots['cores_per_node']
-        loadl_bg_block      = opaque_slots['loadl_bg_block']
-        sub_block_shape_str = opaque_slots['sub_block_shape_str']
-        corner_node         = opaque_slots['corner_node']
-
-        if task_cores % cores_per_node:
-            msg = "Num cores (%d) is not a multiple of %d!" % (task_cores, cores_per_node)
-            self._log.exception(msg)
-            raise ValueError(msg)
-
-        # Runjob it is!
-        runjob_command = self.launch_command
-
-        # Set the number of tasks/ranks per node
-        # TODO: Currently hardcoded, this should be configurable,
-        #       but I don't see how, this would be a leaky abstraction.
-        runjob_command += ' --ranks-per-node %d' % min(cores_per_node, task_cores)
-
-        # Run this subjob in the block communicated by LoadLeveler
-        runjob_command += ' --block %s'  % loadl_bg_block
-        runjob_command += ' --corner %s' % corner_node
-
-        # convert the shape
-        runjob_command += ' --shape %s' % sub_block_shape_str
-
-        # runjob needs the full path to the executable
-        if os.path.basename(task_exec) == task_exec:
-            # Use `which` with back-ticks as the executable,
-            # will be expanded in the shell script.
-            task_exec = '`which %s`' % task_exec
-            # Note: We can't use the expansion from here,
-            #       as the pre-execs of the CU aren't run yet!!
-
-        # And finally add the executable and the arguments
-        # usage: runjob <runjob flags> : /bin/hostname -f
-        runjob_command += ' : %s' % task_exec
-        if task_argstr:
-            runjob_command += ' %s' % task_argstr
-
-        return runjob_command, None
-
-
-# ==============================================================================
-#
-class LaunchMethodDPLACE(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # dplace: job launcher for SGI systems (e.g. on Blacklight)
-        self.launch_command = self._which('dplace')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if 'task_offsets' not in opaque_slots :
-            raise RuntimeError('insufficient information to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        task_offsets = opaque_slots['task_offsets']
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        dplace_offset = task_offsets
-
-        dplace_command = "%s -c %d-%d %s" % (
-            self.launch_command, dplace_offset,
-            dplace_offset+task_cores-1, task_command)
-
-        return dplace_command, None
-
-
-# ==============================================================================
-#
-class LaunchMethodMPIRUNRSH(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        # mpirun_rsh (e.g. on Gordon@SDSC, Stampede@TACC)
-        if not self._which('mpirun_rsh'):
-            raise Exception("mpirun_rsh could not be found")
-
-        # We don't use the full pathname as the user might load a different
-        # compiler / MPI library suite from his CU pre_exec that requires
-        # the launcher from that version, as experienced on stampede in #572.
-        self.launch_command = 'mpirun_rsh'
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if not 'task_slots' in opaque_slots:
-            raise RuntimeError('insufficient information to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        task_slots = opaque_slots['task_slots']
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        # Extract all the hosts from the slots
-        hosts = [slot.split(':')[0] for slot in task_slots]
-
-        # If we have a CU with many cores, we will create a hostfile and pass
-        # that as an argument instead of the individual hosts
-        if len(hosts) > 42:
-
-            # Create a hostfile from the list of hosts
-            hostfile = self._create_hostfile(hosts, impaired=True)
-            hosts_string = "-hostfile %s" % hostfile
-
-        else:
-
-            # Construct the hosts_string ('h1 h2 .. hN')
-            hosts_string = " ".join(hosts)
-
-        export_vars = ' '.join([var+"=$"+var for var in self.EXPORT_ENV_VARIABLES if var in os.environ])
-
-        mpirun_rsh_command = "%s -np %d %s %s %s" % (
-            self.launch_command, task_cores, hosts_string, export_vars, task_command)
-
-        return mpirun_rsh_command, None
-
-
-# ==============================================================================
-#
-class LaunchMethodMPIRUNDPLACE(LaunchMethod):
-    # TODO: This needs both mpirun and dplace
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # dplace: job launcher for SGI systems (e.g. on Blacklight)
-        self.launch_command = self._which('dplace')
-        self.mpirun_command = self._which('mpirun')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if not 'task_offsets' in opaque_slots:
-            raise RuntimeError('insufficient information to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        task_offsets = opaque_slots['task_offsets']
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        dplace_offset = task_offsets
-
-        mpirun_dplace_command = "%s -np %d %s -c %d-%d %s" % \
-            (self.mpirun_command, task_cores, self.launch_command,
-             dplace_offset, dplace_offset+task_cores-1, task_command)
-
-        return mpirun_dplace_command, None
-
-
-
-# ==============================================================================
-#
-class LaunchMethodIBRUN(LaunchMethod):
-    # NOTE: Don't think that with IBRUN it is possible to have
-    # processes != cores ...
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # ibrun: wrapper for mpirun at TACC
-        self.launch_command = self._which('ibrun')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if not 'task_offsets' in opaque_slots:
-            raise RuntimeError('insufficient information to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        task_offsets = opaque_slots['task_offsets']
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        ibrun_offset = task_offsets
-
-        ibrun_command = "%s -n %s -o %d %s" % \
-                        (self.launch_command, task_cores,
-                         ibrun_offset, task_command)
-
-        return ibrun_command, None
-
-
-
-# ==============================================================================
-#
-# NOTE: This requires a development version of Open MPI available.
-#
-class LaunchMethodORTE(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-        # We remove all ORTE related environment variables from the launcher
-        # environment, so that we can use ORTE for both launch of the
-        # (sub-)agent and CU execution.
-        self.env_removables.extend(["OMPI_", "OPAL_", "PMIX_"])
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def lrms_config_hook(cls, name, cfg, lrms, logger):
-        """
-        FIXME: this config hook will manipulate the LRMS nodelist.  Not a nice
-               thing to do, but hey... :P
-               What really should be happening is that the LRMS digs information
-               on node reservation out of the config and configures the node
-               list accordingly.  This config hook should be limited to starting
-               the DVM.
-        """
-
-        dvm_command = cls._which('orte-dvm')
-        if not dvm_command:
-            raise Exception("Couldn't find orte-dvm")
-
-        # Now that we found the orte-dvm, get ORTE version
-        orte_info = {}
-        oi_output = subprocess.check_output(['orte-info|grep "Open RTE"'], shell=True)
-        oi_lines = oi_output.split('\n')
-        for line in oi_lines:
-            if not line:
-                continue
-            key, val = line.split(':')
-            if 'Open RTE' == key.strip():
-                orte_info['version'] = val.strip()
-            elif  'Open RTE repo revision' == key.strip():
-                orte_info['version_detail'] = val.strip()
-        logger.info("Found Open RTE: %s / %s",
-                    orte_info['version'], orte_info['version_detail'])
-
-        # Use (g)stdbuf to disable buffering.
-        # We need this to get the "DVM ready",
-        # without waiting for orte-dvm to complete.
-        # The command seems to be generally available on our Cray's,
-        # if not, we can code some home-coooked pty stuff.
-        stdbuf_cmd =  cls._find_executable(['stdbuf', 'gstdbuf'])
-        if not stdbuf_cmd:
-            raise Exception("Couldn't find (g)stdbuf")
-        stdbuf_arg = "-oL"
-
-        # Base command = (g)stdbuf <args> + orte-dvm + debug_args
-        dvm_args = [stdbuf_cmd, stdbuf_arg, dvm_command]
-
-        # Additional (debug) arguments to orte-dvm
-        debug_strings = [
-            #'--debug-devel',
-            #'--mca odls_base_verbose 100',
-            #'--mca rml_base_verbose 100',
-        ]
-        # Split up the debug strings into args and add them to the dvm_args
-        [dvm_args.extend(ds.split()) for ds in debug_strings]
-
-        vm_size = len(lrms.node_list)
-        logger.info("Starting ORTE DVM on %d nodes with '%s' ...", vm_size, ' '.join(dvm_args))
-        dvm_process = subprocess.Popen(dvm_args, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-
-        dvm_uri = None
-        while True:
-
-            line = dvm_process.stdout.readline().strip()
-
-            if line.startswith('VMURI:'):
-
-                if len(line.split(' ')) != 2:
-                    raise Exception("Unknown VMURI format: %s" % line)
-
-                label, dvm_uri = line.split(' ', 1)
-
-                if label != 'VMURI:':
-                    raise Exception("Unknown VMURI format: %s" % line)
-
-                logger.info("ORTE DVM URI: %s" % dvm_uri)
-
-            elif line == 'DVM ready':
-
-                if not dvm_uri:
-                    raise Exception("VMURI not found!")
-
-                logger.info("ORTE DVM startup successful!")
-                break
-
-            else:
-
-                # Check if the process is still around,
-                # and log output in debug mode.
-                if None == dvm_process.poll():
-                    logger.debug("ORTE: %s" % line)
-                else:
-                    # Process is gone: fatal!
-                    raise Exception("ORTE DVM process disappeared")
-
-        # ----------------------------------------------------------------------
-        def _watch_dvm(dvm_process):
-
-            logger.info('starting DVM watcher')
-
-            while dvm_process.poll() is None:
-                line = dvm_process.stdout.readline().strip()
-                if line:
-                    logger.debug('dvm output: %s' % line)
-                else:
-                    time.sleep(1.0)
-
-            logger.info('DVM stopped (%d)' % dvm_process.returncode)
-            # TODO: Tear down everything?
-        # ----------------------------------------------------------------------
-
-        dvm_watcher = threading.Thread(target=_watch_dvm, args=(dvm_process,),
-                                       name="DVMWatcher")
-        dvm_watcher.daemon = True
-        dvm_watcher.start()
-
-        lm_info = {'dvm_uri'     : dvm_uri,
-                   'version_info': {name: orte_info}}
-
-        # we need to inform the actual LM instance about the DVM URI.  So we
-        # pass it back to the LRMS which will keep it in an 'lm_info', which
-        # will then be passed as part of the opaque_slots via the scheduler
-        return lm_info
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def lrms_shutdown_hook(cls, name, cfg, lrms, lm_info, logger):
-        """
-        This hook is symmetric to the config hook above, and is called during
-        shutdown sequence, for the sake of freeing allocated resources.
-        """
-
-        if 'dvm_uri' in lm_info:
-            try:
-                logger.info('terminating dvm')
-                orte_submit = cls._which('orte-submit')
-                if not orte_submit:
-                    raise Exception("Couldn't find orte-submit")
-                subprocess.Popen([orte_submit, "--hnp", lm_info['dvm_uri'], "--terminate"])
-            except Exception as e:
-                logger.exception('dmv termination failed')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        self.launch_command = self._which('orte-submit')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if 'task_slots' not in opaque_slots:
-            raise RuntimeError('No task_slots to launch via %s: %s' \
-                               % (self.name, opaque_slots))
-
-        if 'lm_info' not in opaque_slots:
-            raise RuntimeError('No lm_info to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        if not opaque_slots['lm_info']:
-            raise RuntimeError('lm_info missing for %s: %s' \
-                               % (self.name, opaque_slots))
-
-        if 'dvm_uri' not in opaque_slots['lm_info']:
-            raise RuntimeError('dvm_uri not in lm_info for %s: %s' \
-                    % (self.name, opaque_slots))
-
-        task_slots = opaque_slots['task_slots']
-        dvm_uri    = opaque_slots['lm_info']['dvm_uri']
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        # Construct the hosts_string, env vars
-        # On some Crays, like on ARCHER, the hostname is "archer_N".
-        # In that case we strip off the part upto and including the underscore.
-        #
-        # TODO: If this ever becomes a problem, i.e. we encounter "real" hostnames
-        #       with underscores in it, or other hostname mangling, we need to turn
-        #       this into a system specific regexp or so.
-        #
-        hosts_string = ",".join([slot.split(':')[0].rsplit('_', 1)[-1] for slot in task_slots])
-        export_vars  = ' '.join(['-x ' + var for var in self.EXPORT_ENV_VARIABLES if var in os.environ])
-
-        # Additional (debug) arguments to orte-submit
-        debug_strings = [
-            #'--debug-devel',
-            #'--mca oob_base_verbose 100',
-            #'--mca rml_base_verbose 100'
-        ]
-        orte_command = '%s %s --hnp "%s" %s -np %s -host %s %s' % (
-            self.launch_command, ' '.join(debug_strings), dvm_uri, export_vars, task_cores, hosts_string, task_command)
-
-        return orte_command, None
-
-
-# ==============================================================================
-#
-class LaunchMethodPOE(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        # poe: LSF specific wrapper for MPI (e.g. yellowstone)
-        self.launch_command = self._which('poe')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_args    = cud.get('arguments') or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        if not 'task_slots' in opaque_slots:
-            raise RuntimeError('insufficient information to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        task_slots = opaque_slots['task_slots']
-
-        # Count slots per host in provided slots description.
-        hosts = {}
-        for slot in task_slots:
-            host = slot.split(':')[0]
-            if host not in hosts:
-                hosts[host] = 1
-            else:
-                hosts[host] += 1
-
-        # Create string with format: "hostX N host
-        hosts_string = ''
-        for host in hosts:
-            hosts_string += '%s %d ' % (host, hosts[host])
-
-        if task_argstr:
-            task_command = "%s %s" % (task_exec, task_argstr)
-        else:
-            task_command = task_exec
-
-        # Override the LSB_MCPU_HOSTS env variable as this is set by
-        # default to the size of the whole pilot.
-        poe_command = 'LSB_MCPU_HOSTS="%s" %s %s' % (
-            hosts_string, self.launch_command, task_command)
-
-        return poe_command, None
-
-
-# ==============================================================================
-#
-# The Launch Method Implementation for Running YARN applications
-#
-class LaunchMethodYARN(LaunchMethod):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg, logger):
-
-        LaunchMethod.__init__(self, cfg, logger)
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def lrms_config_hook(cls, name, cfg, lrms, logger):
-        """
-        FIXME: this config hook will inspect the LRMS nodelist and, if needed,
-               will start the YRN cluster on node[0].
-        """
-
-        logger.info('Hook called by YARN LRMS with the name %s'%lrms.name)
-
-        def config_core_site(node):
-
-            core_site_file = open(os.getcwd()+'/hadoop/etc/hadoop/core-site.xml','r')
-            lines = core_site_file.readlines()
-            core_site_file.close()
-
-            prop_str  = '<property>\n'
-            prop_str += '  <name>fs.default.name</name>\n'
-            prop_str += '    <value>hdfs://%s:54170</value>\n'%node
-            prop_str += '</property>\n'
-
-            lines.insert(-1,prop_str)
-
-            core_site_file = open(os.getcwd()+'/hadoop/etc/hadoop/core-site.xml','w')
-            for line in lines:
-                core_site_file.write(line)
-            core_site_file.close()
-
-        def config_hdfs_site(nodes):
-
-            hdfs_site_file = open(os.getcwd()+'/hadoop/etc/hadoop/hdfs-site.xml','r')
-            lines = hdfs_site_file.readlines()
-            hdfs_site_file.close()
-
-            prop_str  = '<property>\n'
-            prop_str += ' <name>dfs.replication</name>\n'
-            prop_str += ' <value>1</value>\n'
-            prop_str += '</property>\n'
-
-            prop_str += '<property>\n'
-            prop_str += '  <name>dfs.name.dir</name>\n'
-            prop_str += '    <value>file:///tmp/hadoop/hadoopdata/hdfs/namenode</value>\n'
-            prop_str += '</property>\n'
-
-            prop_str += '<property>\n'
-            prop_str += '  <name>dfs.data.dir</name>\n'
-            prop_str += '    <value>file:///tmp/hadoop/hadoopdata/hdfs/datanode</value>\n'
-            prop_str += '</property>\n'
-
-            lines.insert(-1,prop_str)
-
-            hdfs_site_file = open(os.getcwd()+'/hadoop/etc/hadoop/hdfs-site.xml','w')
-            for line in lines:
-                hdfs_site_file.write(line)
-            hdfs_site_file.close()
-
-        def config_mapred_site():
-
-            mapred_site_file = open(os.getcwd()+'/hadoop/etc/hadoop/mapred-site.xml.template','r')
-            lines = mapred_site_file.readlines()
-            mapred_site_file.close()
-
-            prop_str  = ' <property>\n'
-            prop_str += '  <name>mapreduce.framework.name</name>\n'
-            prop_str += '   <value>yarn</value>\n'
-            prop_str += ' </property>\n'
-
-            lines.insert(-1,prop_str)
-
-            mapred_site_file = open(os.getcwd()+'/hadoop/etc/hadoop/mapred-site.xml','w')
-            for line in lines:
-                mapred_site_file.write(line)
-            mapred_site_file.close()
-
-        def config_yarn_site():
-
-            yarn_site_file = open(os.getcwd()+'/hadoop/etc/hadoop/yarn-site.xml','r')
-            lines = yarn_site_file.readlines()
-            yarn_site_file.close()
-
-            prop_str  = ' <property>\n'
-            prop_str += '  <name>yarn.nodemanager.aux-services</name>\n'
-            prop_str += '    <value>mapreduce_shuffle</value>\n'
-            prop_str += ' </property>\n'
-
-            lines.insert(-1,prop_str)
-
-            yarn_site_file = open(os.getcwd()+'/hadoop/etc/hadoop/yarn-site.xml','w')
-            for line in lines:
-                yarn_site_file.write(line)
-            yarn_site_file.close()
-
-        # If the LRMS used is not YARN the namenode url is going to be
-        # the first node in the list and the port is the default one, else
-        # it is the one that the YARN LRMS returns
-        hadoop_home = None
-        if lrms.name == 'YARNLRMS': # FIXME: use constant
-            logger.info('Hook called by YARN LRMS')
-            logger.info('NameNode: {0}'.format(lrms.namenode_url))
-            service_url    = lrms.namenode_url
-            rm_url         = "%s:%s" % (lrms.rm_ip, lrms.rm_port)
-            rm_ip          = lrms.rm_ip
-            launch_command = cls._which('yarn')
-
-        else:
-            # Here are the necessary commands to start the cluster.
-            if lrms.node_list[0] == 'localhost':
-                #Download the tar file
-                node_name = lrms.node_list[0]
-                stat = os.system("wget http://apache.claz.org/hadoop/common/hadoop-2.6.0/hadoop-2.6.0.tar.gz")
-                stat = os.system('tar xzf hadoop-2.6.0.tar.gz;mv hadoop-2.6.0 hadoop;rm -rf hadoop-2.6.0.tar.gz')
-            else:
-                node = subprocess.check_output('/bin/hostname')
-                logger.info('Entered Else creation')
-                node_name = node.split('\n')[0]
-                stat = os.system("wget http://apache.claz.org/hadoop/common/hadoop-2.6.0/hadoop-2.6.0.tar.gz")
-                stat = os.system('tar xzf hadoop-2.6.0.tar.gz;mv hadoop-2.6.0 hadoop;rm -rf hadoop-2.6.0.tar.gz')
-                # TODO: Decide how the agent will get Hadoop tar ball.
-
-                # this was formerly
-                #   def set_env_vars():
-                # but we are in a class method, and don't have self -- and we don't need
-                # it anyway...
-
-            hadoop_home        = os.getcwd() + '/hadoop'
-            hadoop_install     = hadoop_home
-            hadoop_mapred_home = hadoop_home
-            hadoop_common_home = hadoop_home
-            hadoop_hdfs_home   = hadoop_home
-            yarn_home          = hadoop_home
-
-            hadoop_common_lib_native_dir = hadoop_home + '/lib/native'
-
-            #-------------------------------------------------------------------
-            # Solution to find Java's home folder:
-            # http://stackoverflow.com/questions/1117398/java-home-directory
-
-            jpos = subprocess.check_output(['readlink','-f', '/usr/bin/java']).split('bin')
-            if jpos[0].find('jre') != -1:
-                java_home = jpos[0][:jpos[0].find('jre')]
-            else:
-                java_home = jpos[0]
-
-            hadoop_env_file = open(hadoop_home+'/etc/hadoop/hadoop-env.sh','r')
-            hadoop_env_file_lines = hadoop_env_file.readlines()
-            hadoop_env_file.close()
-            hadoop_env_file_lines[24] = 'export JAVA_HOME=%s'%java_home
-            hadoop_env_file = open(hadoop_home+'/etc/hadoop/hadoop-env.sh','w')
-            for line in hadoop_env_file_lines:
-                hadoop_env_file.write(line)
-            hadoop_env_file.close()
-
-            # set_env_vars() ended here
-
-            config_core_site(node_name)
-            config_hdfs_site(lrms.node_list)
-            config_mapred_site()
-            config_yarn_site()
-
-            logger.info('Start Formatting DFS')
-            namenode_format = os.system(hadoop_home + '/bin/hdfs namenode -format -force')
-            logger.info('DFS Formatted. Starting DFS.')
-            hadoop_start = os.system(hadoop_home + '/sbin/start-dfs.sh')
-            logger.info('Starting YARN')
-            yarn_start = os.system(hadoop_home + '/sbin/start-yarn.sh')
-
-            #-------------------------------------------------------------------
-            # Creating user's HDFS home folder
-            logger.debug('Running: %s/bin/hdfs dfs -mkdir /user'%hadoop_home)
-            os.system('%s/bin/hdfs dfs -mkdir /user'%hadoop_home)
-            uname = subprocess.check_output('whoami').split('\n')[0]
-            logger.debug('Running: %s/bin/hdfs dfs -mkdir /user/%s'%(hadoop_home,uname))
-            os.system('%s/bin/hdfs dfs -mkdir /user/%s'%(hadoop_home,uname))
-            check = subprocess.check_output(['%s/bin/hdfs'%hadoop_home,'dfs', '-ls', '/user'])
-            logger.info(check)
-            # FIXME YARN: why was the scheduler configure called here?  Configure
-            #             is already called during scheduler instantiation
-            # self._scheduler._configure()
-
-            service_url = node_name + ':54170'
-            rm_url      = node_name
-            launch_command = yarn_home + '/bin/yarn'
-            rm_ip = node_name
-
-
-        # The LRMS instance is only available here -- everything which is later
-        # needed by the scheduler or launch method is stored in an 'lm_info'
-        # dict.  That lm_info dict will be attached to the scheduler's lrms_info
-        # dict, and will be passed around as part of the opaque_slots structure,
-        # so it is available on all LM create_command calls.
-        lm_info = {'service_url'   : service_url,
-                   'rm_url'        : rm_url,
-                   'hadoop_home'   : hadoop_home,
-                   'rm_ip'         : rm_ip,
-                   'name'          : lrms.name,
-                   'launch_command': launch_command,
-                   'nodename'      : lrms.node_list[0] }
-
-        return lm_info
-
-
-    # --------------------------------------------------------------------------
-    #
-    @classmethod
-    def lrms_shutdown_hook(cls, name, cfg, lrms, lm_info, logger):
-        if 'name' not in lm_info:
-            raise RuntimeError('name not in lm_info for %s' % name)
-
-        if lm_info['name'] != 'YARNLRMS': # FIXME: use constant
-            logger.info('Stoping YARN')
-            os.system(lm_info['hadoop_home'] + '/sbin/stop-yarn.sh')
-
-            logger.info('Stoping DFS.')
-            os.system(lm_info['hadoop_home'] + '/sbin/stop-dfs.sh')
-
-            logger.info("Deleting HADOOP files from temp")
-            os.system('rm -rf /tmp/hadoop*')
-            os.system('rm -rf /tmp/Jetty*')
-            os.system('rm -rf /tmp/hsperf*')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        # Single Node configuration
-        # TODO : Multinode config
-        self._log.info('Getting YARN app')
-        os.system('wget https://dl.dropboxusercontent.com/u/28410803/Pilot-YARN-0.1-jar-with-dependencies.jar')
-        self._log.info(self._cfg['lrms_info']['lm_info'])
-        self.launch_command = self._cfg['lrms_info']['lm_info']['launch_command']
-        self._log.info('YARN was called')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def construct_command(self, cu, launch_script_hop):
-
-        opaque_slots = cu['opaque_slots']
-        work_dir     = cu['workdir']
-        cud          = cu['description']
-        task_exec    = cud['executable']
-        task_cores   = cud['cores']
-        task_env     = cud.get('environment') or {}
-        task_args    = cud.get('arguments')   or []
-        task_argstr  = self._create_arg_string(task_args)
-
-        # Construct the args_string which is the arguments given as input to the
-        # shell script. Needs to be a string
-        self._log.debug("Constructing YARN command")
-        self._log.debug('Opaque Slots {0}'.format(opaque_slots))
-
-        if 'lm_info' not in opaque_slots:
-            raise RuntimeError('No lm_info to launch via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        if not opaque_slots['lm_info']:
-            raise RuntimeError('lm_info missing for %s: %s' \
-                               % (self.name, opaque_slots))
-
-        if 'service_url' not in opaque_slots['lm_info']:
-            raise RuntimeError('service_url not in lm_info for %s: %s' \
-                    % (self.name, opaque_slots))
-
-        if 'rm_url' not in opaque_slots['lm_info']:
-            raise RuntimeError('rm_url not in lm_info for %s: %s' \
-                    % (self.name, opaque_slots))
-
-
-        if 'nodename' not in opaque_slots['lm_info']:
-            raise RuntimeError('nodename not in lm_info for %s: %s' \
-                    % (self.name, opaque_slots))
-
-        service_url = opaque_slots['lm_info']['service_url']
-        rm_url      = opaque_slots['lm_info']['rm_url']
-        client_node = opaque_slots['lm_info']['nodename']
-
-        #-----------------------------------------------------------------------
-        # Create YARN script
-        # This funcion creates the necessary script for the execution of the
-        # CU's workload in a YARN application. The function is responsible
-        # to set all the necessary variables, stage in, stage out and create
-        # the execution command that will run in the distributed shell that
-        # the YARN application provides. There reason for staging out is
-        # because after the YARN application has finished everything will be
-        # deleted.
-
-        print_str ="echo '#!/usr/bin/env bash'>>ExecScript.sh\n"
-        print_str+="echo ''>>ExecScript.sh\n"
-        print_str+="echo ''>>ExecScript.sh\n"
-        print_str+="echo '#---------------------------------------------------------'>>ExecScript.sh\n"
-        print_str+="echo '# Staging Input Files'>>ExecScript.sh\n"
-
-        self._log.debug('Creating input staging')
-        if cud['input_staging']:
-            scp_input_files='"'
-            for InputFile in cud['input_staging']:
-                scp_input_files+='%s/%s '%(work_dir,InputFile['target'])
-            scp_input_files+='"'
-            print_str+="echo 'scp $YarnUser@%s:%s .'>>ExecScript.sh\n"%(client_node,scp_input_files)
-
-        print_str+="echo ''>>ExecScript.sh\n"
-        print_str+="echo ''>>ExecScript.sh\n"
-        print_str+="echo '#---------------------------------------------------------'>>ExecScript.sh\n"
-        print_str+="echo '# Creating Executing Command'>>ExecScript.sh\n"
-        
-        print_str+="echo '%s %s 1>Ystdout 2>Ystderr'>>ExecScript.sh\n"%(cud['executable'],task_argstr)
-
-        print_str+="echo ''>>ExecScript.sh\n"
-        print_str+="echo ''>>ExecScript.sh\n"
-        print_str+="echo '#---------------------------------------------------------'>>ExecScript.sh\n"
-        print_str+="echo '# Staging Output Files'>>ExecScript.sh\n"
-        print_str+="echo 'YarnUser=$(whoami)'>>ExecScript.sh\n"
-        scp_output_files='Ystderr Ystdout'
-
-        if cud['output_staging']:
-            for OutputFile in cud['output_staging']:
-                scp_output_files+=' %s'%(OutputFile['source'])
-        print_str+="echo 'scp -v %s $YarnUser@%s:%s'>>ExecScript.sh\n"%(scp_output_files,client_node,work_dir)
-
-        print_str+="echo ''>>ExecScript.sh\n"
-        print_str+="echo ''>>ExecScript.sh\n"
-        print_str+="echo '#End of File'>>ExecScript.sh\n\n\n"
-
-        env_string = ''
-        for key,val in task_env.iteritems():
-            env_string+= '-shell_env '+key+'='+str(val)+' '
-
-        #app_name = '-appname '+ cud['_id']
-        # Construct the ncores_string which is the number of cores used by the
-        # container to run the script
-        if task_cores:
-            ncores_string = '-container_vcores '+str(task_cores)
-        else:
-            ncores_string = ''
-
-        # Construct the nmem_string which is the size of memory used by the
-        # container to run the script
-        #if task_nummem:
-        #    nmem_string = '-container_memory '+task_nummem
-        #else:
-        #    nmem_string = ''
-
-        #Getting the namenode's address.
-        service_url = 'yarn://%s?fs=hdfs://%s'%(rm_url, service_url)
-
-        yarn_command = '%s -jar ../Pilot-YARN-0.1-jar-with-dependencies.jar'\
-                       ' com.radical.pilot.Client -jar ../Pilot-YARN-0.1-jar-with-dependencies.jar'\
-                       ' -shell_script ExecScript.sh %s %s -service_url %s\ncat Ystdout' % (self.launch_command,
-                        env_string, ncores_string,service_url)
-
-        self._log.debug("Yarn Command %s"%yarn_command)
-
-        return print_str+yarn_command, None
-
 
 # ==============================================================================
 #
@@ -2116,12 +474,12 @@ class AgentExecutingComponent_POPEN (AgentExecutingComponent) :
 
         # The AgentExecutingComponent needs the LaunchMethods to construct
         # commands.
-        self._task_launcher = LaunchMethod.create(
+        self._task_launcher = rp.agent.LM.create(
                 name   = self._cfg.get('task_launch_method'),
                 cfg    = self._cfg,
                 logger = self._log)
 
-        self._mpi_launcher = LaunchMethod.create(
+        self._mpi_launcher = rp.agent.LM.create(
                 name   = self._cfg.get('mpi_launch_method'),
                 cfg    = self._cfg,
                 logger = self._log)
@@ -2555,12 +913,12 @@ class AgentExecutingComponent_SHELL(AgentExecutingComponent):
         # that is the reason of their existence (and opaqueness) in the first
         # place...
 
-        self._task_launcher = LaunchMethod.create(
+        self._task_launcher = rp.agent.LM.create(
                 name   = self._cfg['task_launch_method'],
                 cfg    = self._cfg,
                 logger = self._log)
 
-        self._mpi_launcher = LaunchMethod.create(
+        self._mpi_launcher = rp.agent.LM.create(
                 name   = self._cfg['mpi_launch_method'],
                 cfg    = self._cfg,
                 logger = self._log)
@@ -3126,12 +1484,12 @@ class AgentExecutingComponent_ABDS (AgentExecutingComponent) :
 
         # The AgentExecutingComponent needs the LaunchMethods to construct
         # commands.
-        self._task_launcher = LaunchMethod.create(
+        self._task_launcher = rp.agent.LM.create(
                 name   = self._cfg['task_launch_method'],
                 cfg    = self._cfg,
                 logger = self._log)
 
-        self._mpi_launcher = LaunchMethod.create(
+        self._mpi_launcher = rp.agent.LM.create(
                 name   = self._cfg['mpi_launch_method'],
                 cfg    = self._cfg,
                 logger = self._log)
@@ -4236,7 +2594,7 @@ class AgentWorker(rpu.Worker):
             elif target == 'node':
 
                 if not agent_lm:
-                    agent_lm = LaunchMethod.create(
+                    agent_lm = rp.agent.LM.create(
                         name   = self._cfg['agent_launch_method'],
                         cfg    = self._cfg,
                         logger = self._log)
